view_test: use subquery for finding counts
