+++ dojo/search/views.py
@@ -14,8 +14,7 @@ from dojo.endpoint.queries import get_authorized_endpoints
 from dojo.endpoint.views import prefetch_for_endpoints
 from dojo.engagement.queries import get_authorized_engagements
 from dojo.filters import FindingFilter, FindingFilterWithoutObjectLookups
-from dojo.finding.queries import get_authorized_findings, get_authorized_vulnerability_ids
-from dojo.finding.views import prefetch_for_findings
+from dojo.finding.queries import get_authorized_findings, get_authorized_vulnerability_ids, prefetch_for_findings
 from dojo.forms import SimpleSearchForm
 from dojo.models import Engagement, Finding, Finding_Template, Languages, Product, Test
 from dojo.product.queries import get_authorized_app_analysis, get_authorized_products
